Moved LGDevice to seperate class
